+++ Source~/DangFugBixs.Generators~/DangFugBixs.Sandbox/Scopes/LifetimeScopes.cs
@@ -23,7 +23,7 @@ public class GameLifetimeScope : LifetimeScope {
 /// <summary>
 /// A test scope mapping to an Identity Type in a different assembly (Runtime.dll)
 /// </summary>
-[NhemDangFugBixs.Attributes.LifetimeScopeFor<NhemDangFugBixs.Runtime.Testing.CrossLayerIdentity>]
+[NhemDangFugBixs.Attributes.LifetimeScopeFor(typeof(NhemDangFugBixs.Runtime.Testing.CrossLayerIdentity))]
 public class CrossLayerLifetimeScope : LifetimeScope {
 }
 
@@ -41,7 +41,7 @@ public class GameplayLifetimeScope : LifetimeScope {
 /// UI-specific lifetime scope with custom name override.
 /// </summary>
 [ScopeName("UI")]
-[NhemDangFugBixs.Attributes.LifetimeScopeFor<NhemDangFugBixs.Runtime.Testing.CrossLayerIdentity>] // Just for testing multiple mappings
+[NhemDangFugBixs.Attributes.LifetimeScopeFor(typeof(NhemDangFugBixs.Runtime.Testing.CrossLayerIdentity))] // Just for testing multiple mappings
 public class UserInterfaceLifetimeScope : LifetimeScope {
     protected override void Configure(IContainerBuilder builder) {
         NhemDangFugBixs.Generated.DangFugBixsSandbox.VContainerRegistration.RegisterUI(builder);